api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asarPy==1.0.1
click==8.3.2
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
